refactor: reorganize file structure
